+++ src/main/java/com/medicdefense/backend/payment/infrastructure/persistence/jpa/ConsultationRepository.java
@@ -2,5 +2,5 @@ package com.medicdefense.backend.payment.infrastructure.persistence.jpa;
 import com.medicdefense.backend.payment.domain.model.aggregates.Consultation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ConsultationRepository extends JpaRepository<Consultation, Integer> {
+public interface ConsultationRepository extends JpaRepository<Consultation, Long> {
 }
